requirements.txt
attrs==20.3.0
flake8==3.8.4
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
toml==0.10.2
